Added front-end language cast additions
Added new type dyn and cast t2 in t1 : exp expressions to SML.
With this language, we can finally write examples to test the
intermediate language!
Next stage is to add cast behaviour into the Lambda intermediate
language. For now, the behaviour is dummy (just compile the inner
expression and discard the casts).